do not overwrite user's configuration unexpectedly
previous code is executed only once when the file is loaded. So it may
be executed in unintended file and does not work when user executes
`git rebase -i` repeatedly.

`gitrebase` is a filetype which is set to rebase buffer. So using
`autocmd` ensures to execute the `setlocal` command for each git rebase
buffers.